requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
google-genai>=1.0.0
python-docx>=1.1.0
